chore: restructure UserInput and fix string states type
